requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2
python-dotenv==1.0.0
supabase>=2.8.0
openai==0.27.8
python-multipart==0.0.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.95.0
